+++ .vscode/settings.json
@@ -73,6 +73,7 @@
         "textinput",
         "Toggleable",
         "toposort",
+        "unlocalized",
         "usemap",
         "valign",
         "vlist",
