+++ packages-ui/roosterjs-react/lib/common/type/DialogItem.ts
@@ -1,0 +1,9 @@
+/**
+ * @internal
+ */
+export default interface DialogItem<Strings extends string> {
+    labelKey: Strings | null;
+    unlocalizedLabel: string;
+    initValue: string;
+    autoFocus?: boolean;
+}
